fix: removed unused components
